feat: type guard 생성, type 안정성 강화
